+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 1.5.0 IN PROGRESS
 
+* Adjust checkout columns width. Fixes UICHKOUT-457.
+
 
 ## [1.4.0](https://github.com/folio-org/ui-checkout/tree/v1.4.0) (2018-12-13)
 [Full Changelog](https://github.com/folio-org/ui-checkout/compare/v1.3.1...v1.4.0)