+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 1.5.0 IN PROGRESS
 
 * Adjust checkout columns width. Fixes UICHKOUT-457.
+* Fix enforce manual patron blocks. Fixes UIU-792.
 
 
 ## [1.4.0](https://github.com/folio-org/ui-checkout/tree/v1.4.0) (2018-12-13)